refactor: rename chartView to chart-view and cardView to card-view
